% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_add_limits.R
\name{chron_add_limits}
\alias{chron_add_limits}
\title{Add limits of each chronology control table}
\usage{
chron_add_limits(data_source)
}
\arguments{
\item{data_source}{Data.frame with \code{chron_control_format}}
}
\description{
Add limits of each chronology control table
}
