% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_add_postbomb_curve.R
\name{chron_add_postbomb_curve}
\alias{chron_add_postbomb_curve}
\title{Change calibration curve to postbomb if necessary}
\usage{
chron_add_postbomb_curve(
  data_source,
  post_bomb_cal_curve,
  postbomb_age = 199,
  rc_control_types = NULL
)
}
\arguments{
\item{data_source}{Chronology control table}

\item{post_bomb_cal_curve}{Name of the postbomb curve to use}

\item{postbomb_age}{Threshold old for use of the postbomb curve}

\item{rc_control_types}{Vector with all RC types}
}
\value{
Chronology control table with corrected calibration curves
}
\description{
Change calibration curve to postbomb if necessary, based on
type of chronology control point types and `postbomb_age``
}
