% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_change_pmc_to_ages.R
\name{chron_change_pmc_to_ages}
\alias{chron_change_pmc_to_ages}
\title{Back-transform percent modern carbon to C14 ages}
\usage{
chron_change_pmc_to_ages(data_source)
}
\arguments{
\item{data_source}{Chronology control table}
}
\value{
Return chronology control table with corrected ages
}
\description{
Back-transform percent modern carbon percentage using
IntCal package to real radiocarbon ages which will likely be negative and
fall in the future. These needs a follow step calibrating with postbomb
calibration curves
}
