% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_check_negative_rc_ages.R
\name{chron_check_negative_rc_ages}
\alias{chron_check_negative_rc_ages}
\title{Check if there are any radiocarbon ages with negative values}
\usage{
chron_check_negative_rc_ages(
  data_source,
  postbomb_age = 199,
  rc_control_types = NULL
)
}
\arguments{
\item{data_source}{Chronology control table}

\item{postbomb_age}{Threshold old for use of the postbomb curve}

\item{rc_control_types}{Vector with all RC types}
}
\value{
Logical
}
\description{
Detect radiocarbon ages with negative values
}
\keyword{internal}
