% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_decimal_ages_detect.R
\name{chron_decimal_ages_detect}
\alias{chron_decimal_ages_detect}
\title{Test if chronology control table has ages with decimal places}
\usage{
chron_decimal_ages_detect(data_source, rc_control_types = NULL)
}
\arguments{
\item{data_source}{Chronology control table}

\item{rc_control_types}{Vector with all radiocarbon types}
}
\value{
Logical
}
\description{
Test if chronology control table has a potential candidate
of Percentage modern carbon
}
\keyword{internal}
