% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_decimal_ages_mark.R
\name{chron_decimal_ages_mark}
\alias{chron_decimal_ages_mark}
\title{Mark chronology control points, which have ages with decimal places}
\usage{
chron_decimal_ages_mark(
  data_source,
  only_rc = TRUE,
  rc_control_types = NULL,
  age_range = c(80, 120)
)
}
\arguments{
\item{data_source}{Chronology control table}

\item{only_rc}{Should only radiocarbon control points be tested?}

\item{rc_control_types}{Vector with all radiocarbon types}

\item{age_range}{The age period to search for decimal places. Default is
80-120 as decimal places occur around 100.}
}
\value{
The original data.frame with a new column \code{fulfil_criteria} which is
marking ages with place
}
\description{
Detect ages or radiocarbon control points, which have
decimal places are are between 80 and 120.
}
\keyword{internal}
