% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_fix_pmc.R
\name{chron_fix_pmc}
\alias{chron_fix_pmc}
\title{Detect potential chronology control points saved as Percentage
modern carbon and let user decide, which should be back-transformed}
\usage{
chron_fix_pmc(data_source, chron_control_types, dir)
}
\arguments{
\item{data_source}{Data.frame with \code{dataset_id}, \code{chron_control}, and
\code{chron_control_format}}

\item{chron_control_types}{Named list with \code{chroncontrol_included_types} and
\code{radiocarbon_control_types}}

\item{dir}{Path to the data storage folder}
}
\description{
Detect potential chronology control points saved as Percentage
modern carbon and let user decide, which should be back-transformed
}
\keyword{internal}
