% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_fix_postbomb.R
\name{chron_fix_postbomb}
\alias{chron_fix_postbomb}
\title{Change calibration curves to postbomb if needed in all chronology
control tables}
\usage{
chron_fix_postbomb(data_source, chron_control_types, set_postbomb_age = 199)
}
\arguments{
\item{data_source}{Data.frame with \code{dataset_id}, \code{chron_control_format}, and
\code{postbomb_curve_name}}

\item{chron_control_types}{Named list with \code{chroncontrol_included_types} and
\code{radiocarbon_control_types}}

\item{set_postbomb_age}{Age determining the start of period using postbomb
curves. Default value is  199 as all postbomb curves are able to handle
values younger than 199}
}
\description{
Change calibration curves to postbomb if needed in all chronology
control tables
}
