% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_format_table.R
\name{chron_format_table}
\alias{chron_format_table}
\title{Transform the chronology control into Bchron format and filter
the points}
\usage{
chron_format_table(
  data_source,
  curve_name,
  use_default_thickness,
  use_default_error,
  max_error,
  valid_control_types,
  control_types_to_calibrate,
  guess_depth
)
}
\arguments{
\item{data_source}{Chronology control tables}

\item{curve_name}{Type of calibration curve assign to the whole site}

\item{use_default_thickness}{Value to use if thickness is not present}

\item{use_default_error}{Value to use if error is not present}

\item{max_error}{Maximum error value to accept}

\item{valid_control_types}{List of valid control types}

\item{control_types_to_calibrate}{List of control types to use calibration curve}

\item{guess_depth}{Maximum depth to be accepted "Guess" as valid type}
}
\value{
Data.frame with chronology control table
}
\description{
Will automatically prepare the chron control table checking
and adjusting various properties. For example, duplication of depth
missing values, calibration curve specifics, etc
}
