% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_get_control_tables.R
\name{chron_get_control_tables}
\alias{chron_get_control_tables}
\title{Extract chronology control tables from data assembly}
\usage{
chron_get_control_tables(data_source, min_n_of_control_points = 2)
}
\arguments{
\item{data_source}{Data.frame with \code{dataset_id},\code{region}, \code{chron_control},
\code{n_chron_control}, \code{curve_name}, and\code{postbomb_curve_name}}

\item{min_n_of_control_points}{Minimal number of chronology control points
each table has to have to be included. Minimal is 2}
}
\description{
Extract chronology control tables from data assembly
}
