% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_confirm_current_state.R
\name{chron_get_current_state}
\alias{chron_get_current_state}
\title{The current chronology setting and state of previous results}
\usage{
chron_get_current_state(dir, calc_AD_models_denovo, predict_ages_denovo)
}
\arguments{
\item{dir}{Path to the data storage folder}

\item{calc_AD_models_denovo}{Setting to define if age-depth models should be
re-calibrated 'de novo'}

\item{predict_ages_denovo}{Setting to define if ages should be predicted from
last models 'de novo'}
}
\value{
Named list containing the summary of the presence of previous results and current settings
}
\description{
The current chronology setting and state of previous results
}
