% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_get_limits.R
\name{chron_get_limits}
\alias{chron_get_limits}
\title{Detect the youngest and oldest age of chronology control points}
\usage{
chron_get_limits(data_source)
}
\arguments{
\item{data_source}{Chronology control table}
}
\value{
Vector with youngest and oldest.
}
\description{
Detect the youngest and oldest age of chronology control points
}
