% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_ggsave_bchron_output.R
\name{chron_ggsave_bchron_output}
\alias{chron_ggsave_bchron_output}
\title{Save chronology as ggplot}
\usage{
chron_ggsave_bchron_output(
  data_source,
  dataset_id,
  dir,
  text_size,
  line_size,
  image_width,
  image_height,
  image_units = c("in", "cm", "mm", "px"),
  image_format = ".pdf"
)
}
\arguments{
\item{data_source}{Chronology data}

\item{dataset_id}{dataset_id}

\item{dir}{Directory path}

\item{text_size}{Text size}

\item{line_size}{Line size}

\item{image_width}{Width of image}

\item{image_height}{Height of image}

\item{image_units}{Units to measure image}

\item{image_format}{Format of figure}
}
\description{
Save chronology as ggplot
}
