% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_load_data_for_ad_modelling.R
\name{chron_load_data_for_ad_modelling}
\alias{chron_load_data_for_ad_modelling}
\title{Load chronology control tables and filter our those present in Crash
File}
\usage{
chron_load_data_for_ad_modelling(dir)
}
\arguments{
\item{dir}{Path to the data storage folder}
}
\description{
Load chronology control tables and filter our those present in Crash
File
}
