% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_merge_results.R
\name{chron_merge_results}
\alias{chron_merge_results}
\title{Load and merge processed data with their predicted ages}
\usage{
chron_merge_results(dir)
}
\arguments{
\item{dir}{Path to the data storage folder}
}
\description{
Load and merge processed data with their predicted ages
}
