% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_predict_all_ages.R
\name{chron_predict_all_ages}
\alias{chron_predict_all_ages}
\title{Predict ages for the selected data assembly}
\usage{
chron_predict_all_ages(data_source, dir, date)
}
\arguments{
\item{data_source}{Data.frame with \code{dataset_id} and \code{bchron_mod}}

\item{dir}{Path to the data storage folder}

\item{date}{Date to named the files with}
}
\value{
Data with \code{levels}, which contains \code{depth} and predicted \verb{age, and }age_uncertainty`, which contains age uncertainty matrix.
}
\description{
Load the depth information from the last saved data assembly,
then predict the ages using the age-depth models in \code{data_source}.
}
