% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_prepare_ad_to_predict.R
\name{chron_prepare_ad_to_predict}
\alias{chron_prepare_ad_to_predict}
\title{Prepare data assembly of sequences to predict aged for based on the
presence of previous results and the user-defined settings}
\usage{
chron_prepare_ad_to_predict(
  data_source,
  dir,
  sites_to_rerun = c(),
  current_state
)
}
\arguments{
\item{data_source}{Data.frame containing the \code{dataset_id}}

\item{dir}{Path to the data storage folder}

\item{sites_to_rerun}{Vector with \code{dataset_id}s of sequences to rerun even
if previous results are present}

\item{current_state}{Named list congaing the summary of the presence of
previous results and current settings}
}
\description{
Prepare data assembly of sequences to predict aged for based on the
presence of previous results and the user-defined settings
}
