% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_prepare_cal_curves.R
\name{chron_prepare_cal_curves}
\alias{chron_prepare_cal_curves}
\title{Extract calibration curves and copy them into Bchron package}
\usage{
chron_prepare_cal_curves(first_curve_proportion = 0.5)
}
\arguments{
\item{first_curve_proportion}{The proportion of contribution of the IntCal curve. Default is 0.5}
}
\description{
Extract calibration curves and copy them into Bchron package
}
