% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_prepare_control_tables.R
\name{chron_prepare_control_tables}
\alias{chron_prepare_control_tables}
\title{Format chronology control tables and filter out tables with specific
number of control points}
\usage{
chron_prepare_control_tables(
  data_source,
  chron_control_types,
  default_thickness,
  default_error,
  max_age_error,
  guess_depth,
  min_n_of_control_points = 2
)
}
\arguments{
\item{data_source}{Data.frame containing \code{chron_control} and \code{curve_name}}

\item{chron_control_types}{Named list with \code{chroncontrol_included_types} and
\code{radiocarbon_control_types}}

\item{default_thickness}{Value to use if thickness is not present}

\item{default_error}{Value to use if error is not present}

\item{max_age_error}{Maximum error value to accept}

\item{guess_depth}{Maximum depth to be accepted "Guess" as valid type}

\item{min_n_of_control_points}{Minimal number of chronology control points}
}
\description{
Format chronology control tables and filter out tables with specific
number of control points
}
\keyword{internal}
