% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_prepare_for_ad_modelling.R
\name{chron_prepare_for_ad_modelling}
\alias{chron_prepare_for_ad_modelling}
\title{Prepare all chronology control tables for age-depth modelling}
\usage{
chron_prepare_for_ad_modelling(
  data_source,
  chron_control_types,
  default_thickness,
  default_error,
  max_age_error,
  guess_depth,
  min_n_of_control_points = 2,
  dir
)
}
\arguments{
\item{data_source}{Data.frame containing \code{chron_control} and \code{curve_name}}

\item{chron_control_types}{Named list with \code{chroncontrol_included_types} and
\code{radiocarbon_control_types}}

\item{default_thickness}{value to use if thickness is not present}

\item{default_error}{value to use if error is not present}

\item{max_age_error}{maximum error value to accept}

\item{guess_depth}{maximum depth to be accepted "Guess" as valid type}

\item{min_n_of_control_points}{Minimal number of chronology control points
each table has to have to be included. Minimal is 2}

\item{dir}{Path to the data storage folder}
}
\description{
Format the chronology control tables, adjust postbomb-ages (if
necessary), adjust Percentage Modern Carbon (if necessary), detect age limits
of chronology control tables, output only variables needed for age-depth
modelling
}
