% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_recalibrate_ad_models.R
\name{chron_recalibrate_ad_models}
\alias{chron_recalibrate_ad_models}
\title{Recalibrate all age-depth models base on chronology control tables}
\usage{
chron_recalibrate_ad_models(
  data_source,
  batch_size = 1,
  number_of_cores = 1,
  default_iteration = 10000,
  default_burn = 2000,
  default_thin = 8,
  iteration_multiplier = 5,
  set_seed = 1234,
  batch_attemps = 3,
  dir
)
}
\arguments{
\item{data_source}{Data.frame containing \code{dataset_id} and \code{chron_control_format}}

\item{batch_size}{Number of individual sequences re-calibrate in a single
batch using parallel computation}

\item{number_of_cores}{Number of CPU cores to use in parallel computation}

\item{default_iteration}{The number of iterations used by Bchron}

\item{default_burn}{The number of starting iterations to discard used by Bchron}

\item{default_thin}{The step size for every iteration to keep beyond
the burnin used by Bchron}

\item{iteration_multiplier}{Value to be used to multiply \code{iteration}, \code{burn},
\code{thin} to keep the ratio between them same}

\item{set_seed}{User-defined seed for randomisations}

\item{batch_attemps}{Number of tries each batch should be considered
before skipping it}

\item{dir}{Path to the data storage folder}
}
\description{
Recalibrate all age-depth models base on chronology control tables
}
