% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_recalibrate_in_batches.R
\name{chron_recalibrate_in_batches}
\alias{chron_recalibrate_in_batches}
\title{Recalibrate the selected sequences in defined batches using parallel
computation}
\usage{
chron_recalibrate_in_batches(
  data_source_chron,
  data_source_batch,
  dir,
  n_iterations = 10000,
  n_burn = 2000,
  n_thin = 8,
  number_of_cores = 1,
  set_seed = 1234,
  maximum_number_of_loops = 3
)
}
\arguments{
\item{data_source_chron}{Data.frame containing \code{dataset_id} and \code{chron_control_format}}

\item{data_source_batch}{Data.frame with the description of the status
of individual batches}

\item{dir}{Path to the data storage folder}

\item{n_iterations}{The number of iterations used by Bchron}

\item{n_burn}{The number of starting iterations to discard used by Bchron}

\item{n_thin}{The step size for every iteration to keep beyond
the burnin used by Bchron}

\item{number_of_cores}{Number of CPU cores to use in parallel computation}

\item{set_seed}{User-defined seed for randomisations}

\item{maximum_number_of_loops}{Number of tries each batch should be considered
before skipping it}
}
\value{
Named list with \code{bchron_output}, \code{batch_success_table}
}
\description{
Recalibrate the selected sequences in defined batches using parallel
computation
}
