% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_run_bchron.R
\name{chron_run_bchron}
\alias{chron_run_bchron}
\title{Run Bchron age-depth model estimation}
\usage{
chron_run_bchron(data_source, n_iterations = 10000, n_burn = 2000, n_thin = 8)
}
\arguments{
\item{data_source}{Chronology control table}

\item{n_iterations}{The number of iterations to run}

\item{n_burn}{The number of starting iterations to discard}

\item{n_thin}{The step size for every iteration to keep beyond the burnin}
}
\value{
output \code{Bchron} class object
}
\description{
Try to run Bchron. If not successful, return NA
}
