% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_save_ad_figures.R
\name{chron_save_ad_figures}
\alias{chron_save_ad_figures}
\title{Save all age depth models by regions}
\usage{
chron_save_ad_figures(
  dir,
  date,
  text_size,
  line_size,
  image_width,
  image_height,
  image_units = c("in", "cm", "mm", "px"),
  image_format = ".pdf"
)
}
\arguments{
\item{dir}{Path to the data_storage folder}

\item{date}{Date to be used for making folders}

\item{text_size}{Text size}

\item{line_size}{Line size}

\item{image_width}{Width of image}

\item{image_height}{Height of image}

\item{image_units}{Units to measure image}

\item{image_format}{Format of figure}
}
\description{
Load the most recent data, create a folder with current data
and save all plots in it by regions
}
