% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_save_results.R
\name{chron_save_results}
\alias{chron_save_results}
\title{Save outputs of the chronology estimation as age-depth models and
predicted ages}
\usage{
chron_save_results(
  data_source_predicted_ages,
  data_source_to_predict,
  current_state,
  sites_to_rerun,
  dir
)
}
\arguments{
\item{data_source_predicted_ages}{Data.frame with \code{dataset_id} and \code{bchron_mod}}

\item{data_source_to_predict}{Data.frame with \code{dataset_id}, \code{levels}, and
\code{age_uncertainty}}

\item{current_state}{Named list congaing the summary of the presence of
previous results and current settings}

\item{sites_to_rerun}{Vector with \code{dataset_id}s of sequences to rerun even
if previous results are present}

\item{dir}{Path to the data storage folder}
}
\description{
Save outputs of the chronology estimation as age-depth models and
predicted ages
}
