% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_add_data<-,proj_db_class-method}
\alias{db_add_data<-,proj_db_class-method}
\title{Add data to the database}
\usage{
\S4method{db_add_data}{proj_db_class}(theObject) <- value
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
New data will be automatically added to the specific tables
but duplicates will be removed
}
