% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Affiliations,proj_db_class-method}
\alias{db_Affiliations,proj_db_class-method}
\title{Get the Affiliation table}
\usage{
\S4method{db_Affiliations}{proj_db_class}(theObject)
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
Get the Affiliation table
}
