% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Affiliations<-,proj_db_class-method}
\alias{db_Affiliations<-,proj_db_class-method}
\title{Insert in the Affiliation table}
\usage{
\S4method{db_Affiliations}{proj_db_class}(theObject) <- value
}
\arguments{
\item{theObject}{object of proj_db_class class}

\item{value}{object to assign}
}
\description{
New data will be added but duplicates will be removed
}
