% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Affiliations<-}
\alias{db_Affiliations<-}
\title{Insert in the Affiliation table}
\usage{
db_Affiliations(theObject) <- value
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
Insert in the Affiliation table
}
