% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Affiliations}
\alias{db_Affiliations}
\title{Get the Affiliation table}
\usage{
db_Affiliations(theObject)
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
Get the Affiliation table
}
