% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Auth_aff_tab<-}
\alias{db_Auth_aff_tab<-}
\title{Insert in table linking Authors and Affiliations}
\usage{
db_Auth_aff_tab(theObject) <- value
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
Insert in table linking Authors and Affiliations
}
