% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Auth_dataset_tab<-}
\alias{db_Auth_dataset_tab<-}
\title{Insert in the table linking Authors and Datasets}
\usage{
db_Auth_dataset_tab(theObject) <- value
}
\arguments{
\item{theObject}{object of proj_db_class class}

\item{value}{object to assign}
}
\description{
Insert in the table linking Authors and Datasets
}
