% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Authors<-,proj_db_class-method}
\alias{db_Authors<-,proj_db_class-method}
\title{Insert in Author table}
\usage{
\S4method{db_Authors}{proj_db_class}(theObject) <- value
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
New data will be added but duplicates will be removed
}
