% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Authors<-}
\alias{db_Authors<-}
\title{Insert in Author table}
\usage{
db_Authors(theObject) <- value
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
Insert in Author table
}
