% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Dataset_pub<-,proj_db_class-method}
\alias{db_Dataset_pub<-,proj_db_class-method}
\title{Insert in the table with publication linked to datasets}
\usage{
\S4method{db_Dataset_pub}{proj_db_class}(theObject) <- value
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
New data will be added but duplicates will be removed
}
