% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Dataset_pub}
\alias{db_Dataset_pub}
\title{Get the table with publication linked to datasets}
\usage{
db_Dataset_pub(theObject)
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
Get the table with publication linked to datasets
}
