% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{db_Datasets_id_tab}
\alias{db_Datasets_id_tab}
\title{Get the table with datasets information}
\usage{
db_Datasets_id_tab(theObject)
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
Get the table with datasets information
}
