% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{default_show}
\alias{default_show}
\title{Function show summary of the database}
\usage{
default_show(theObject)
}
\arguments{
\item{theObject}{object of proj_db_class class}
}
\description{
Function show summary of the database
}
