% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dataset_from_excel.R
\name{extract_dataset_from_excel}
\alias{extract_dataset_from_excel}
\title{Extract the selected site into Neotoma-style data.frame}
\usage{
extract_dataset_from_excel(selected_file, dir, suffix = "", project_db)
}
\arguments{
\item{selected_file}{A name of the site to import in quotes}

\item{dir}{Directory path}

\item{suffix}{A character, which would add to all new IDs
(dataset, authors, affiliations, etc).}

\item{project_db}{File including the project dataset database}
}
\value{
Neotoma-style data.frame
}
\description{
Import the dataset in excel format, check if dataset
is unique and assign unique author and affiliation IDs, add count data,
chronology control points, and publication information, and make data.frame
consistent with Neotoma-processed style data.
}
