% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_assign_shapefile.R
\name{geo_assign_shapefile}
\alias{geo_assign_shapefile}
\title{Extract data from shapefile and add to points}
\usage{
geo_assign_shapefile(data_source, shapefile)
}
\arguments{
\item{data_source}{Data.frame containing \code{long} and \code{lat}}

\item{shapefile}{Shapefile to use}
}
\value{
Data.frame with additional columns
}
\description{
Extract data from shapefile and add to points provided based
on the data \code{lat} and \code{long}
}
