% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_assign_tif.R
\name{geo_assign_tif}
\alias{geo_assign_tif}
\title{Extract data from TIF file and add as points}
\usage{
geo_assign_tif(
  data_source,
  tif_file_name = "",
  fill_na = FALSE,
  na_as_value = NULL,
  distance_step = 500,
  n_max_step = 10
)
}
\arguments{
\item{data_source}{Data.frame containing \code{long} and \code{lat}}

\item{tif_file_name}{Name of a shape file to use}

\item{fill_na}{Logical. If TRUE, function will search for the most common value
in the increasing distance of 'distance_step'}

\item{na_as_value}{Use for values of tiff file, which should be treated as
NA for the purposes of looking for additional value.}

\item{distance_step}{Numeric. Distance in meters giving radius to increase the
search radius}

\item{n_max_step}{Numeric. Maximum number of step increases.}
}
\value{
Data.frame with additional columns
}
\description{
Extract data from tif file and add to points provided based
on the data \code{lat} and \code{long}. If \code{fill_na} == TRUE, function will search for
the most common value in the increasing distance of \code{distance_step}.
Function will stop search if all values are found.
}
