% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_assign_value.R
\name{geo_assign_value}
\alias{geo_assign_value}
\title{Assign geographical value to a dataset}
\usage{
geo_assign_value(
  data_source,
  dir,
  sel_method = c("shapefile", "tif"),
  file_name,
  var = "raster_values",
  var_name,
  tif_fill_na = TRUE,
  tif_distance_step = 5000,
  tif_n_max_step = 20
)
}
\arguments{
\item{data_source}{Data.frame with \code{lat} and \code{long} coordinates to be assign a
value}

\item{dir}{Directory of the source of information}

\item{sel_method}{What is the format of the source of information.
Could be either \code{shapefile} or \code{tif}}

\item{file_name}{Name of the layer for shapefile or tif file}

\item{var}{Which variable should be extracted from source of information?
Note that it is set as "raster_values" as it is default for tifs}

\item{var_name}{Optional name of the new variable for extracted data}

\item{tif_fill_na}{Logical. For tif only. If \code{TRUE}, function will search
for most common value in the increasing distance of \code{tif_distance_step}}

\item{tif_distance_step}{For tif only. Use for values, which should be
treated as \code{NA} for the purposes of looking for additional value.}

\item{tif_n_max_step}{Numeric. For tif only. Distance in m giving radius
to increase the search radius}
}
\description{
Extract data from shape file or tif and add to points
provided in \code{data_source} based on the data \code{lat} and \code{long}. There are
additional options for tif files. If \code{tif_fill_na == TRUE}, function will
search for the most common value in the increasing distance of
\code{tif_distance_step}. Function will stop search if all values are found.
}
