% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harm_get_all_tables.R
\name{harm_get_all_tables}
\alias{harm_get_all_tables}
\title{Get all harmonisation tables for regions present}
\usage{
harm_get_all_tables(data_source, dir)
}
\arguments{
\item{data_source}{Data.frame with \code{harmonisation_region} and \code{raw_counts}}

\item{dir}{Path to the data storage folder}
}
\description{
Get all harmonisation tables for regions present
}
