% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_all_regions.R
\name{harmonise_all_regions}
\alias{harmonise_all_regions}
\title{Harmonise each region by its harmonisation table}
\usage{
harmonise_all_regions(
  data_source,
  harmonisation_tables,
  original_name = "taxon_name",
  harm_level = "level_1",
  exclude_taxa = "delete",
  pollen_grain_test = TRUE
)
}
\arguments{
\item{data_source}{Data.frame with \code{dataset_id}, \code{harmonisation_region}, and
\code{raw_counts}}

\item{harmonisation_tables}{Data.frame with \code{harmonisation_region} and
\code{harm_table}}

\item{original_name}{Character. Name of the column in harmonisation table
to be used.}

\item{harm_level}{Name of the column to use for harmonisation}

\item{exclude_taxa}{Character. Name of the taxa, which should be omitted.}

\item{pollen_grain_test}{Logical. Test for difference in sum of total
pollen sum.}
}
\description{
Harmonise each region by its harmonisation table
}
