% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_taxa.R
\name{harmonise_taxa}
\alias{harmonise_taxa}
\title{Harmonise the taxa to a certain taxonomic level}
\usage{
harmonise_taxa(
  data_source,
  harmonisation_table,
  original_name = "taxon_name",
  harm_name = "level_1",
  exclude_taxa = "delete",
  pollen_grain_test = TRUE
)
}
\arguments{
\item{data_source}{Table containing the raw pollen count}

\item{harmonisation_table}{Table including taxon original names
(\code{original_name}) as well as harmonised (\code{harm_name})}

\item{original_name}{Character. Name of the column in harmonisation table
of original counts}

\item{harm_name}{Character. Name of the column in harmonisation table
to be used.}

\item{exclude_taxa}{Character. Name of the taxa, which should be omitted.}

\item{pollen_grain_test}{Logical. Test for difference in sum of total
pollen sum before and after harmonisation.}
}
\value{
Data.frame with harmonised counts
}
\description{
Harmonise the taxa to a certain taxonomic level
}
