% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_datasets_from_folder.R
\name{import_datasets_from_folder}
\alias{import_datasets_from_folder}
\title{Import all datasets from selected directory}
\usage{
import_datasets_from_folder(dir_files, dir, suffix = "private")
}
\arguments{
\item{dir_files}{A directory path to the folder with all Excel sheets}

\item{dir}{Path to data storage folder}

\item{suffix}{A character, which would add to all new IDs
(dataset, authors, affiliations, etc).}
}
\value{
Data.frame with all data from the folder formatted in a
Neotoma-processed style data.
}
\description{
The function will source all datasets from selected \code{dir_files}
folder and compare it will all datasets in project database (object
of class \code{proj_db_class}) from \code{dir}. It will automatically
update the database if needed and use information from the database in
case that the dataset/author/affiliation is already present.
}
