% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_lenght_of_data.R
\name{plot_age_lenght_of_data}
\alias{plot_age_lenght_of_data}
\title{Plot age-limits of all records in dataset compilation}
\usage{
plot_age_lenght_of_data(
  data_source,
  line_width = 1,
  line_alpha = 1,
  line_colour = "gray30",
  text_size = 16,
  line_size = 0.1,
  legend_position = c("none", "bottom", "top", "left", "right")
)
}
\arguments{
\item{data_source}{Data.frame to containing the dataset compilation}

\item{line_width}{Numeric. Width of a line representing one record}

\item{line_alpha}{Numeric. Alpha of each line}

\item{line_colour}{Character. Color of each line. Can be also a name of
a variable in \code{data_source}}

\item{text_size}{Numeric. Text size for axis}

\item{line_size}{Numeric. Width of a line for axis}

\item{legend_position}{Character. If legend present, where it should be
placed? Default is "none", which will not display legend.
See \code{legend.position} in \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\description{
Plot age-limits of all records in dataset compilation
}
\seealso{
\code{\link[=plot_graphical_summary]{plot_graphical_summary()}}, \code{\link[=plot_map_of_data]{plot_map_of_data()}},
\code{\link[=plot_count_of_data]{plot_count_of_data()}},
}
