% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count_of_data.R
\name{plot_count_of_data}
\alias{plot_count_of_data}
\title{A bar plot with number of records in dataset compilation}
\usage{
plot_count_of_data(
  data_source,
  point_color = "gray80",
  point_size = 10,
  ouline_colour = "gray30",
  line_size = 0.1,
  text_size = 16,
  legend_position = c("none", "bottom", "top", "left", "right")
)
}
\arguments{
\item{data_source}{Data.frame to containing the dataset compilation}

\item{point_color}{Character. Color of lollipops. Can be also a name of
a variable in \code{data_source}}

\item{point_size}{Numeric. Size of a lollipop}

\item{ouline_colour}{Numeric. Colour of a outline of lollipol}

\item{line_size}{Numeric. Width of a line for axis}

\item{text_size}{Numeric. Text size for axis}

\item{legend_position}{Character. If legend present, where it should be
placed? Default is "none", which will not display legend.
See \code{legend.position} in \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\description{
A bar plot with number of records in dataset compilation
}
\seealso{
\code{\link[=plot_graphical_summary]{plot_graphical_summary()}}, \code{\link[=plot_map_of_data]{plot_map_of_data()}},
\code{\link[=plot_age_lenght_of_data]{plot_age_lenght_of_data()}}
}
