% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_of_data.R
\name{plot_map_of_data}
\alias{plot_map_of_data}
\title{Plot map of geographical location of data compilation}
\usage{
plot_map_of_data(
  data_source,
  point_size = 3,
  point_alpha = 1,
  point_colour = "gray30",
  point_colour_accent = NULL,
  col_map_fill = "gray90",
  col_map_borders = NA,
  map_data_margin = 1,
  text_size = 16,
  line_size = 0.1,
  legend_position = c("none", "bottom", "top", "left", "right")
)
}
\arguments{
\item{data_source}{Data.frame to containing the dataset compilation}

\item{point_size}{Numeric. Size of a point representing one record}

\item{point_alpha}{Numeric. Alpha of each point}

\item{point_colour}{Character. Color of each point. Can be also a name of
a variable in \code{data_source}}

\item{point_colour_accent}{Optional. Character. If present, each point will
be highlighted by such color.}

\item{col_map_fill}{Character. A color of a land in map}

\item{col_map_borders}{Character. A color of a country borders in map.
\code{NA} will produce no border lines}

\item{map_data_margin}{Numeric. Number of degrees set as a margin aroud data
to be ploted}

\item{text_size}{Numeric. Text size for axis}

\item{line_size}{Numeric. Line width}

\item{legend_position}{Character. If legend present, where it should be
placed? Default is "none", which will not display legend.
See \code{legend.position} in \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\description{
Plot map of geographical location of data compilation
}
\seealso{
\code{\link[=plot_graphical_summary]{plot_graphical_summary()}}, \code{\link[=plot_age_lenght_of_data]{plot_age_lenght_of_data()}},
\code{\link[=plot_count_of_data]{plot_count_of_data()}},
}
