% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pollen_diagram_rioja.R
\name{plot_pollen_diagram_rioja}
\alias{plot_pollen_diagram_rioja}
\title{Plot a pollen diagram using the rioja package}
\usage{
plot_pollen_diagram_rioja(
  data_percentages,
  levels,
  y_var = c("age", "depth"),
  dataset_id,
  min_n_occur = 3,
  max_taxa = 20,
  dir
)
}
\arguments{
\item{data_percentages}{Data.frame pollen percentages}

\item{levels}{Data.frame with level information}

\item{y_var}{The name of the variable to be used as Y-axis in the chart}

\item{dataset_id}{Unique ID of the sequences}

\item{min_n_occur}{Min occurrence of taxa to be displayed}

\item{max_taxa}{Number of taxa to plotted per one page}

\item{dir}{Folder to save the figure to}
}
\description{
Plot a pollen diagram using the \code{rioja} package
}
