% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_add_region_age_limits.R
\name{proc_add_region_age_limits}
\alias{proc_add_region_age_limits}
\title{Define age limits for each region in the data}
\usage{
proc_add_region_age_limits(data_source, dir)
}
\arguments{
\item{data_source}{Data.frame with \code{region} variable}

\item{dir}{Directory of data storage folder}
}
\description{
Define age limits for each region in the data
}
