% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_clean_count_names.R
\name{proc_clean_count_names}
\alias{proc_clean_count_names}
\title{Clean names for count datasets}
\usage{
proc_clean_count_names(data_source, additional_patterns = NULL)
}
\arguments{
\item{data_source}{Data.frame containing \code{raw_counts} column which contains
raw counts}

\item{additional_patterns}{User defined patterns so be used for name changes}
}
\description{
Clean names for count datasets
}
