% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_clean_names.R
\name{proc_clean_names}
\alias{proc_clean_names}
\title{Clean the names of the columns}
\usage{
proc_clean_names(data_source, additional_patterns = NULL)
}
\arguments{
\item{data_source}{Vector with names (characters)}

\item{additional_patterns}{User defined patterns so be used for name changes}
}
\value{
Vector with new names
}
\description{
First, change special characters to text and then
use janitor package to clean the rest.
}
