% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_detect_age_limits.R
\name{proc_detect_age_limits}
\alias{proc_detect_age_limits}
\title{Test if sequence is long enough for selected age criteria}
\usage{
proc_detect_age_limits(
  data_source,
  age_limit_young,
  age_limit_old,
  test_quantiles = FALSE
)
}
\arguments{
\item{data_source}{Data.frame with level ages}

\item{age_limit_young}{Young age limit for selected site}

\item{age_limit_old}{Old age limit for selected site}

\item{test_quantiles}{Test quantiles of age prediction?}
}
\value{
Logical
}
\description{
Test if the youngest limit of any level is younger than
age limit as well as if the oldest limit of any level is older than age
limit
}
