% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_detect_rowsum_distribution.R
\name{proc_detect_rowsum_distribution}
\alias{proc_detect_rowsum_distribution}
\title{Test if enough levels fulfil criteria of pollen sum}
\usage{
proc_detect_rowsum_distribution(
  data_counts,
  data_levels,
  age_limit_young,
  age_limit_old,
  target_n_grains,
  percentage_samples,
  test_quantiles = FALSE
)
}
\arguments{
\item{data_counts}{Data.frame with pollen data}

\item{data_levels}{Data.frame with level ages}

\item{age_limit_young}{young age limit for selected site}

\item{age_limit_old}{old age limit for selected site}

\item{target_n_grains}{preferred number of pollen sum}

\item{percentage_samples}{percentages of levels, which are in the age limit
to be tested if their pollen sum fulfil criteria}

\item{test_quantiles}{Test quantiles of age prediction?}
}
\value{
TRUE/FALSE (logical)
}
\description{
Calculate pollen sum of each level and test if it is above
target_n_grains. Compare record has at least X\% (percentage_samples) of
levels, within the age period, with enough pollen sum. Level is counted
"in" period if any part of his 95th age quantile is between age_limit_young
and age_limit_old.
}
