% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_filter_all_data.R
\name{proc_filter_all_data}
\alias{proc_filter_all_data}
\title{Filter out levels and records based on the criteria}
\usage{
proc_filter_all_data(
  data_source,
  msg = NA_character_,
  variable_vec = c("levels", "raw_counts", "counts_harmonised", "age_uncertainty"),
  filter_by_pollen_sum = TRUE,
  min_n_grains = 0,
  target_n_grains = 100,
  percentage_samples = 0,
  filter_by_age_limit = TRUE,
  filter_by_extrapolation = TRUE,
  maximum_age_extrapolation = Inf,
  filter_by_interest_region = TRUE,
  filter_by_number_of_levels = TRUE,
  min_n_levels = TRUE,
  use_age_quantiles = FALSE,
  use_bookend_level = FALSE
)
}
\arguments{
\item{data_source}{Data.frame with  \code{dataset_id}.}

\item{msg}{A message to be outputted when error occur with age limits}

\item{variable_vec}{Vector with name sof columns which should be filtered.
By default the columns are \code{levels}, \code{raw_counts}, \code{counts_harmonised}, and
\code{age_uncertainty}}

\item{filter_by_pollen_sum}{Logical. If true, levels and records will be
filtered out by \code{min_n_grains}, \code{target_n_grains}, \code{percentage_samples}}

\item{min_n_grains}{Number of individual pollen grains which each level has
to have}

\item{target_n_grains}{Number of individual pollen grains which each levels
'preferably' has to have}

\item{percentage_samples}{Threshold of number of levels with 'preferable' counts}

\item{filter_by_age_limit}{Logical. If true, records will be filtered out
based on their age limits, defined by \code{young_age} and \code{old_age}}

\item{filter_by_extrapolation}{Logical. If true, records will be filtered
out based on the age of the last chronology control point}

\item{maximum_age_extrapolation}{Maximum age, which be can be extrapolated
beyond the oldest chronology control point}

\item{filter_by_interest_region}{Logical. If true, filter out levels beyond
the age of interest}

\item{filter_by_number_of_levels}{Logical. If true, filter out records based
on the number of levels}

\item{min_n_levels}{Minimal number of levels each record has to have}

\item{use_age_quantiles}{Logical. Should 95th age quantile be used for data
filtration? This will result in more stable data assembly between different
result of AD modelling BUT require additional data preparation before
analytical part}

\item{use_bookend_level}{Logical. Should all data filtration omit one
additional level in the old period? This will result of "bookend" level,
which can help to provide anchor information after the period of interest}
}
\description{
Filter out levels and records based on the criteria
}
