% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_filter_by_dep_env.R
\name{proc_filter_by_dep_env}
\alias{proc_filter_by_dep_env}
\title{Filter data by the depositional environments}
\usage{
proc_filter_by_dep_env(
  data_source,
  selection_data,
  data_storage_path,
  dir_spec
)
}
\arguments{
\item{data_source}{Data.frame with records to be filtered}

\item{selection_data}{Data.frame with depositional environments presented in the data}

\item{data_storage_path}{Path to the data storage folder}

\item{dir_spec}{Specification of the folder name}
}
\description{
Use "stop-check" (see \code{stopcheck_table} function) to load/create
the table defined by the user. Next, it will filter the records in
\code{data_source} based on the selected depositional environments.
}
