% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_filter_by_geography.R
\name{proc_filter_by_geography}
\alias{proc_filter_by_geography}
\title{Filter data by geographical information}
\usage{
proc_filter_by_geography(
  data_source,
  long_min = NA,
  long_max = NA,
  lat_min = NA,
  lat_max = NA,
  alt_min = NA,
  alt_max = NA
)
}
\arguments{
\item{data_source}{Data.frame with geographical information to be filtered out}

\item{long_min}{Limit for the smallest longitude}

\item{long_max}{Limit for the largest longitude}

\item{lat_min}{Limit for the smallest latitude}

\item{lat_max}{Limit for the largest latitude}

\item{alt_min}{Limit for the smallest altitude}

\item{alt_max}{Limit for the largest altitude}
}
\description{
Filter data by geographical information
}
