% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_filter_by_min.R
\name{proc_filter_by_min}
\alias{proc_filter_by_min}
\title{Filter records by minimal value}
\usage{
proc_filter_by_min(data_source, var_name, min_n)
}
\arguments{
\item{data_source}{Data.frame to be filtered}

\item{var_name}{Name of the variable to be compared to}

\item{min_n}{Criterion to be evaluated against}
}
\description{
Filter records by minimal value
}
\keyword{internal}
