% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_filter_out_duplicates.R
\name{proc_filter_out_duplicates}
\alias{proc_filter_out_duplicates}
\title{Filter out duplicated dataset}
\usage{
proc_filter_out_duplicates(
  data_source,
  data_storage_path,
  filter_var,
  n_subgroups = NULL,
  max_degree_distance = 1
)
}
\arguments{
\item{data_source}{Data frame containing the data to be filtered}

\item{data_storage_path}{Path to the data storage folder}

\item{filter_var}{Name of a column that indicates which column to be used to
compare the data (character)}

\item{n_subgroups}{number of subgroups to split the data based on the
geography (numeric). Sites within subgroup will be tested only if subgroup
contains sites form different \code{filter_var}}

\item{max_degree_distance}{maximal euclidean distance between two sites to
consider them as material for comparison (numeric)}
}
\description{
Filter out duplicated dataset
}
