% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_get_sampleid_age_lim.R
\name{proc_get_sampleid_age_lim}
\alias{proc_get_sampleid_age_lim}
\title{Select levels within the focal time period}
\usage{
proc_get_sampleid_age_lim(
  data_source,
  age_limit,
  current_year_cuttof = TRUE,
  current_year_cuttof_value = -76,
  test_quantiles = FALSE,
  bookend = FALSE
)
}
\arguments{
\item{data_source}{Data.frame with ages of levels}

\item{age_limit}{Maximum age value, marking the end of interest period for
selected site}

\item{current_year_cuttof}{should cut off levels be beyond current year?}

\item{current_year_cuttof_value}{year value of current year}

\item{test_quantiles}{Test quantiles of age prediction?}

\item{bookend}{Keep one extra level beyond the limit?}
}
\value{
Vector with valid sample_ids
}
\description{
Exclude levels which are beyond the \code{age_limit}
User can use age quantiles instead of set ages using \code{test_quantiles} or
keep one level beyond the age extra i.e. \code{bookend}
}
