% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_get_sampleid_extrapol.R
\name{proc_get_sampleid_extrapol}
\alias{proc_get_sampleid_extrapol}
\title{Test if levels are within the limit of age extrapolation}
\usage{
proc_get_sampleid_extrapol(
  data_level,
  data_chron_control_limits,
  maximum_age_extrapolation,
  test_quantiles = FALSE
)
}
\arguments{
\item{data_level}{Data.frame with ages of levels}

\item{data_chron_control_limits}{Vector of two (min and max age of control
points)}

\item{maximum_age_extrapolation}{number of the maximum allowed
extrapolation}

\item{test_quantiles}{Test quantiles of age prediction?}
}
\value{
Vector with valid \code{sample_ids}
}
\description{
Select only levels, which has the age (our quantiles)
younger than last chron.control point + \code{maximum_age_extrapolation} &
older than first chron.control point - \code{maximum_age_extrapolation}
}
\keyword{internal}
