% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_get_sampleid_rowsums.R
\name{proc_get_sampleid_rowsums}
\alias{proc_get_sampleid_rowsums}
\title{Select levels by the total pollen sum}
\usage{
proc_get_sampleid_rowsums(data_source, min_n_grains)
}
\arguments{
\item{data_source}{Data.frame with pollen data}

\item{min_n_grains}{A number of pollen grains}
}
\value{
Vector with valid sample_ids
}
\description{
Calculates pollen sum and test if each level has more pollen
sum than \code{min_n_grains}.
}
\keyword{internal}
