% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_add_chronologies.R
\name{proc_neo_add_chronologies}
\alias{proc_neo_add_chronologies}
\title{Add chronologies to the meta table}
\usage{
proc_neo_add_chronologies(neotoma_meta_samples, chroncontrol_tables)
}
\arguments{
\item{neotoma_meta_samples}{Data.frame, which contains the meta information}

\item{chroncontrol_tables}{Chronology tables extracted form Neotoma}
}
\description{
Add chronologies to the meta table
}
