% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_check_final_assembly.R
\name{proc_neo_check_final_assembly}
\alias{proc_neo_check_final_assembly}
\title{Check and adjust samples for sample and raw count data}
\usage{
proc_neo_check_final_assembly(data_source)
}
\arguments{
\item{data_source}{Data.frame with \code{sample_depth} and \code{raw_counts} information}
}
\description{
Check and adjust samples for sample and raw count data
}
