% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_download_records.R
\name{proc_neo_download_records}
\alias{proc_neo_download_records}
\title{Download all records from Neotoma}
\usage{
proc_neo_download_records(allds, dir, n_tries = 10)
}
\arguments{
\item{allds}{Data.frame with all dataset IDs}

\item{dir}{Path to the data storage folder}

\item{n_tries}{Number of tries to download each dataset}
}
\value{
List with all downloaded records
}
\description{
Download all records from Neotoma
}
