% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_extract_counts.R
\name{proc_neo_extract_counts}
\alias{proc_neo_extract_counts}
\title{Extract pollen data from nested Neotoma data and filter it}
\usage{
proc_neo_extract_counts(data_source, sel_var_element = "pollen", sel_eco_group)
}
\arguments{
\item{data_source}{Data.frame with nested \code{sample_detail}}

\item{sel_var_element}{Character with selected variable element}

\item{sel_eco_group}{Vector with selected ecological groups}
}
\value{
Data.frame for each of the dataset with the pollen counts
}
\description{
Will create data set with count data that should
contain pollen types within \code{sel_eco_group}. The default provide
the counts for the pollen data.
}
