% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_extract_samples.R
\name{proc_neo_extract_samples}
\alias{proc_neo_extract_samples}
\title{Extract sample data from neotoma2 structure}
\usage{
proc_neo_extract_samples(data_source, sel_dataset_id)
}
\arguments{
\item{data_source}{List of sites (neotoma2 JSON structure)}

\item{sel_dataset_id}{Character with selected variable.element}
}
\value{
Data.frame for each of the datasets with nested sample data
}
\description{
Create a table with one row, that include \code{dataset_id}
and nested table with all the sample info.
}
