% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_authors.R
\name{proc_neo_get_authors}
\alias{proc_neo_get_authors}
\title{Obtain information about authors of datasets}
\usage{
proc_neo_get_authors(neotoma_download, dataset_ids, dir, download_new = TRUE)
}
\arguments{
\item{neotoma_download}{List of lists with Neotoma data}

\item{dataset_ids}{List of all dataset_id to extract contact information}

\item{dir}{Path to the data storage folder}

\item{download_new}{Logical. Download the full info about authors? Time
consuming}
}
\description{
Obtain information about authors of datasets
}
