% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_chronologies.R
\name{proc_neo_get_chronologies}
\alias{proc_neo_get_chronologies}
\title{Obtain chronology control tables from Neotoma download}
\usage{
proc_neo_get_chronologies(
  neotoma_download,
  chron_order,
  min_n_of_control_points
)
}
\arguments{
\item{neotoma_download}{List of lists with Neotoma data}

\item{chron_order}{Data.frame with preferred order of chronology types}

\item{min_n_of_control_points}{Minimal number of chronology control points in
selected chronology table}
}
\description{
Obtain chronology control tables from Neotoma download
}
