% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_coord.R
\name{proc_neo_get_coord}
\alias{proc_neo_get_coord}
\title{Get coordinates of sequences from Neotoma list}
\usage{
proc_neo_get_coord(datasets)
}
\arguments{
\item{datasets}{List of lists with Neotoma datasets}
}
\description{
Get coordinates of sequences from Neotoma list
}
