% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_dataset_id.R
\name{proc_neo_get_dataset_id}
\alias{proc_neo_get_dataset_id}
\title{Get all \code{dataset_id} from Neotoma download}
\usage{
proc_neo_get_dataset_id(neotoma_download)
}
\arguments{
\item{neotoma_download}{List of lists with Neotoma data}
}
\description{
Get all \code{dataset_id} from Neotoma download
}
\keyword{internal}
