% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_dep_env.R
\name{proc_neo_get_dep_env}
\alias{proc_neo_get_dep_env}
\title{Obtain the table of depositional environments from Neotoma}
\usage{
proc_neo_get_dep_env()
}
\description{
Obtain the table of depositional environments from Neotoma
}
