% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_metadata.R
\name{proc_neo_get_metadata}
\alias{proc_neo_get_metadata}
\title{Obtain meta information from Neotoma download}
\usage{
proc_neo_get_metadata(neotoma_download)
}
\arguments{
\item{neotoma_download}{List of lists with Neotoma data}
}
\value{
Data.frame with meta information such as site names, coordinates, doi, etc
}
\description{
Obtain meta information from Neotoma download
}
