% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_metadata_with_samples.R
\name{proc_neo_get_metadata_with_samples}
\alias{proc_neo_get_metadata_with_samples}
\title{Obtain meta and sample information from Neotoma download}
\usage{
proc_neo_get_metadata_with_samples(
  neotoma_download,
  long_min = NA,
  long_max = NA,
  lat_min = NA,
  lat_max = NA,
  alt_min = NA,
  alt_max = NA
)
}
\arguments{
\item{neotoma_download}{List of lists with Neotoma data}

\item{long_min}{Limit for the smallest longitude}

\item{long_max}{Limit for the largest longitude}

\item{lat_min}{Limit for the smallest latitude}

\item{lat_max}{Limit for the largest latitude}

\item{alt_min}{Limit for the smallest altitude}

\item{alt_max}{Limit for the largest altitude}
}
\description{
Obtain meta and sample information from Neotoma download
}
