% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_raw_counts.R
\name{proc_neo_get_raw_counts}
\alias{proc_neo_get_raw_counts}
\title{Obtain the counts and filter them by ecological groups}
\usage{
proc_neo_get_raw_counts(data_source, sel_var_element = "pollen")
}
\arguments{
\item{data_source}{Data.frame with samples saved as nested information}

\item{sel_var_element}{Name of the element to selected the raw counts (default
is 'pollen')}
}
\description{
Obtain the counts and filter them by ecological groups
}
