% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_sample_depth.R
\name{proc_neo_get_sample_depth}
\alias{proc_neo_get_sample_depth}
\title{Get sample depth data from samples}
\usage{
proc_neo_get_sample_depth(data_source, min_n_levels = 1)
}
\arguments{
\item{data_source}{Data.frame with samples saved as nested information}

\item{min_n_levels}{Minimal number of levels for records to be included}
}
\description{
Get sample depth data from samples
}
