% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_samples.R
\name{proc_neo_get_samples}
\alias{proc_neo_get_samples}
\title{Obtain sample information from Neotoma download}
\usage{
proc_neo_get_samples(neotoma_download)
}
\arguments{
\item{neotoma_download}{List of lists with Neotoma data}
}
\value{
Data.farme with nested information about individual samples
}
\description{
Obtain sample information from Neotoma download
}
