% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_samples_eco_group.R
\name{proc_neo_get_samples_eco_group}
\alias{proc_neo_get_samples_eco_group}
\title{Extract all ecological groups from dataset}
\usage{
proc_neo_get_samples_eco_group(data_source)
}
\arguments{
\item{data_source}{Data.frame with samples saved as nested information}
}
\description{
Extract all ecological groups from dataset
}
