% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_sites.R
\name{proc_neo_get_sites}
\alias{proc_neo_get_sites}
\title{Obtain sites information from Neotoma download}
\usage{
proc_neo_get_sites(neotoma_download)
}
\arguments{
\item{neotoma_download}{List of lists with Neotoma data}
}
\value{
List of lists with all site information
}
\description{
Obtain sites information from Neotoma download
}
