% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_update_project_database.R
\name{proc_neo_update_project_database}
\alias{proc_neo_update_project_database}
\title{Update the project database with information from Neotoma}
\usage{
proc_neo_update_project_database(
  neotoma_sites_authors,
  neotoma_meta_samples,
  dir
)
}
\arguments{
\item{neotoma_sites_authors}{Data.frame with extracted information about
authors form Neotoma}

\item{neotoma_meta_samples}{Data.frame with datasets present in data
compilation}

\item{dir}{Path to the data storage folder}
}
\description{
Update the project database with information from Neotoma
}
