% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_other_filter_by_dep_env.R
\name{proc_other_filter_by_dep_env}
\alias{proc_other_filter_by_dep_env}
\title{Filter Other sequences by the selected depositional environments}
\usage{
proc_other_filter_by_dep_env(data_source, data_storage_path)
}
\arguments{
\item{data_source}{Data.frame with sequences and their depositional environments}

\item{data_storage_path}{Path to the data storage folder}
}
\description{
The function will obtain the all depositional environments from
Neotoma. Next, it will filter the sequences based on the selected
depositional environments.
}
