% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_other_prepare.R
\name{proc_other_prepare}
\alias{proc_other_prepare}
\title{Prepare other datasets}
\usage{
proc_other_prepare(
  data_source,
  data_storage_path,
  min_n_levels = 1,
  long_min = NA,
  long_max = NA,
  lat_min = NA,
  lat_max = NA,
  alt_min = NA,
  alt_max = NA
)
}
\arguments{
\item{data_source}{Data.frame containing other data assembly}

\item{data_storage_path}{Path to the data storage folder}

\item{min_n_levels}{Minimal number of levels for sequences to be included}

\item{long_min}{Limit for the smallest longitude}

\item{long_max}{Limit for the largest longitude}

\item{lat_min}{Limit for the smallest latitude}

\item{lat_max}{Limit for the largest latitude}

\item{alt_min}{Limit for the smallest altitude}

\item{alt_max}{Limit for the largest altitude}
}
\description{
Data is checked for names. Filtered by depositional
environments, geographical location, and number of samples
}
