% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_prepare_raw_count_levels.R
\name{proc_prepare_raw_count_levels}
\alias{proc_prepare_raw_count_levels}
\title{Format count and depth data}
\usage{
proc_prepare_raw_count_levels(data_source)
}
\arguments{
\item{data_source}{Data.frame with \code{raw_counts} and \code{sample_depth}}
}
\description{
\code{sample_depth} Data.frame will be sorted by \code{depth} and then
\code{raw_counts} data.frame levels will be subset and ordered by the \code{sample_id}
}
