% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rename_column.R
\name{proc_rename_column}
\alias{proc_rename_column}
\title{Change name of a columns id data.frame}
\usage{
proc_rename_column(data_source, old_name, new_name)
}
\arguments{
\item{data_source}{Data.frame to fix}

\item{old_name}{Character. old name of column}

\item{new_name}{Character. new name of column}
}
\value{
Data.frame with new name
}
\description{
Detect if table has a column with \code{old_name} and replace
it with \code{new_name}
}
