% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_save_assembly.R
\name{proc_save_assembly}
\alias{proc_save_assembly}
\title{Save the final data assemblage}
\usage{
proc_save_assembly(
  data_source,
  select_final_variables,
  user_sel_variables = c(),
  dir
)
}
\arguments{
\item{data_source}{Data.frame with \code{dataset_id} and other variables}

\item{select_final_variables}{Logical. Should the final variables be selected
by interactively via R console? If FALSE, all variables will be selected}

\item{user_sel_variables}{Vector with variables, which have to present in the
final data assembly}

\item{dir}{Path to the data storage folder}
}
\description{
Save the final data assemblage
}
