% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_subset_all_data_by_id.R
\name{proc_subset_all_data_by_id}
\alias{proc_subset_all_data_by_id}
\title{Subset data assemblage by valid \code{sample_id}}
\usage{
proc_subset_all_data_by_id(
  data_source,
  variable_vec = c("levels", "raw_counts", "counts_harmonised", "age_uncertainty")
)
}
\arguments{
\item{data_source}{Data assembly. Table containing variables defined
by \code{variable_vec}and \code{valid_id}.}

\item{variable_vec}{Vector with names of columns which should be filtered}
}
\value{
Data assembly will all level-related columns filtered by \code{valid_id}
}
\description{
Subset data assemblage by valid \code{sample_id}
}
