% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_subset_levels.R
\name{proc_subset_levels}
\alias{proc_subset_levels}
\title{Subset levels in data.frame by the vector of \code{sample_ids}}
\usage{
proc_subset_levels(data_source, level_vector)
}
\arguments{
\item{data_source}{Data.frame including \code{sample_id}}

\item{level_vector}{Vector of \code{sample_ids}}
}
\value{
Data.frame only including the levels in \code{level_vector}
}
\description{
Subset levels in data.frame by the vector of \code{sample_ids}
}
