% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_subset_uncertainty_matrix.R
\name{proc_subset_uncertainty_matrix}
\alias{proc_subset_uncertainty_matrix}
\title{Subset levels in data.frame by the vector of \code{sample_ids}}
\usage{
proc_subset_uncertainty_matrix(data_source, level_vector)
}
\arguments{
\item{data_source}{Matrix including \code{sample_id} as columns}

\item{level_vector}{Vector of \code{sample_ids}}
}
\value{
Matrix only including the columns in \code{level_vector}
}
\description{
Subset levels in data.frame by the vector of \code{sample_ids}
}
\keyword{internal}
