% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\docType{class}
\name{proj_db_class-class}
\alias{proj_db_class-class}
\alias{proj_db_class}
\title{Project dataset database}
\description{
Project dataset database
}
\section{Slots}{

\describe{
\item{\code{Authors}}{data.frame with information about authors}

\item{\code{Affiliations}}{data.frame with affiliation details}

\item{\code{Auth_aff_tab}}{data.frame linking authors and affiliations}

\item{\code{Datasets_id_tab}}{data.frame with information about individual datasets}

\item{\code{Auth_dataset_tab}}{data.frame linking authors and datasets}

\item{\code{Dataset_pub}}{data.frame with publications for datasets}
}}

