% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossilpol-package.R
\docType{package}
\name{RFossilpol-package}
\alias{RFossilpol}
\alias{RFossilpol-package}
\title{RFossilpol: Tools for FOSSILPOL workflow}
\description{
This package provides tools for FOSSILPOL workflow, which process global palaeoecological pollen data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hope-uib-bio.github.io/FOSSILPOL-website/}
}

}
\author{
\strong{Maintainer}: Ondrej Mottl \email{ondrej.mottl@gmail.com} (\href{https://orcid.org/0000-0002-9796-5081}{ORCID})

}
\keyword{internal}
