% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFossilpol-package.R
\docType{package}
\name{RFossilpol}
\alias{RFossilpol}
\alias{RFossilpol-package}
\title{RFossilpol: R-package to provide tools for FOSSILPOL workflow}
\description{
The RFossilpol package is to provide tools for FOSSILPOL workflow. See more information about in on dedicated FOSSILPOL website (https://hope-uib-bio.github.io/FOSSILPOL-website/)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hope-uib-bio.github.io/FOSSILPOL-website/}
}

}
\author{
\strong{Maintainer}: Ondrej Mottl \email{ondrej.mottl@gmail.com} (\href{https://orcid.org/0000-0002-9796-5081}{ORCID})

}
\keyword{internal}
