% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_database_class.R
\name{show,proj_db_class-method}
\alias{show,proj_db_class-method}
\title{Show summary of the database}
\usage{
\S4method{show}{proj_db_class}(object)
}
\arguments{
\item{object}{object of proj_db_class class}
}
\description{
Show summary of the database
}
