% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopcheck_make_age_limits.R
\name{stopcheck_make_age_limits}
\alias{stopcheck_make_age_limits}
\title{Create file with age limits for regions}
\usage{
stopcheck_make_age_limits(data_source, dir)
}
\arguments{
\item{data_source}{Data.frame with the regions}

\item{dir}{Path to save file}
}
\value{
Data.frame including age limits for all regions
}
\description{
Create file with age limits for regions
}
