% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopcheck_make_chron_types.R
\name{stopcheck_make_chron_types}
\alias{stopcheck_make_chron_types}
\title{Create file with list of chronology control point types}
\usage{
stopcheck_make_chron_types(data_source, dir)
}
\arguments{
\item{data_source}{Data.frame with the chronology control types}

\item{dir}{Path to save file}
}
\value{
Data.frame including all chronology control points
}
\description{
Create file with list of chronology control point types
}
\keyword{internal}
