% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopcheck_make_default.R
\name{stopcheck_make_default}
\alias{stopcheck_make_default}
\title{Create file with list of all present data}
\usage{
stopcheck_make_default(data_source, dir)
}
\arguments{
\item{data_source}{Data.frame with the data}

\item{dir}{Path to save file}
}
\value{
Data.frame including all present data and \code{include} == FALSE
}
\description{
Create file with list of all present data
}
\keyword{internal}
