% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopcheck_make_harmonisation.R
\name{stopcheck_make_harmonisation}
\alias{stopcheck_make_harmonisation}
\title{Create file with list of all taxa to be harmonised}
\usage{
stopcheck_make_harmonisation(data_source, dir)
}
\arguments{
\item{data_source}{Data.frame with the \code{raw_counts} data}

\item{dir}{Path to save file}
}
\value{
Data.frame with all taxa
}
\description{
Create file with list of all taxa to be harmonised
}
