% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopcheck_make_potent_dupl.R
\name{stopcheck_make_potent_dupl}
\alias{stopcheck_make_potent_dupl}
\title{Create file with list of potential duplicates}
\usage{
stopcheck_make_potent_dupl(data_source, dir)
}
\arguments{
\item{data_source}{Data.frame with the potential duplicates}

\item{dir}{Path to save file}
}
\value{
Data.frame including all potential duplicates
}
\description{
Create file with list of potential duplicates
}
\keyword{internal}
