% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopcheck_table.R
\name{stopcheck_table}
\alias{stopcheck_table}
\title{Check presence of table and create new one if needed}
\usage{
stopcheck_table(
  data_source,
  file_name,
  dir,
  sel_method = c("default", "chron_control", "harm_table", "duplicates", "age_limits"),
  msg = "",
  stop_session = TRUE
)
}
\arguments{
\item{data_source}{Data.frame containing the data to use}

\item{file_name}{Name of the file to load/create}

\item{dir}{Path to the file storage}

\item{sel_method}{Method to select appropriate function to create a table}

\item{msg}{Message to output as a guide to alter the table}

\item{stop_session}{Logical. If true, the session will be stopped}
}
\value{
Object from \code{file_name} from \code{dir}
}
\description{
Function check the presence of the selected \code{file_name} in
\code{dir}. If there is a file, function will load the table and return it.
If there is not a file, function will create the table using function
selected based on \code{sel_method} and stop script
}
