% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_data_table.R
\name{util_check_data_table}
\alias{util_check_data_table}
\title{Check the result of filtration process on table}
\usage{
util_check_data_table(data_source, msg = "")
}
\arguments{
\item{data_source}{Data.frame of sequences}

\item{msg}{Message to be appended to the end of warning}
}
\description{
Function will check the number of rows of the table and return
warning message \code{msg} if there are no sequences
}
