% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_if_loaded.R
\name{util_check_if_loaded}
\alias{util_check_if_loaded}
\title{Confirm that the selected file is present in selected environment}
\usage{
util_check_if_loaded(file_name, env = rlang::current_env, silent = FALSE)
}
\arguments{
\item{file_name}{Name of the selected file}

\item{env}{The environment to test the presence of the file in}

\item{silent}{Logical. Should there be no message?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
