% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_presence_of_lastest_chronology.R
\name{util_check_presence_of_lastest_chronology}
\alias{util_check_presence_of_lastest_chronology}
\title{Check the presence of previous results of chronology estimation}
\usage{
util_check_presence_of_lastest_chronology(dir)
}
\arguments{
\item{dir}{Path to the data storage folder}
}
\value{
Named vector with a summary of presence of previous chronologies and
predicted ages
}
\description{
Check the presence of previous results of chronology estimation
}
