% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_the_latest_file.R
\name{util_check_the_latest_file}
\alias{util_check_the_latest_file}
\title{Detect the newest version of the selected file}
\usage{
util_check_the_latest_file(file_name, dir, folder = FALSE)
}
\arguments{
\item{file_name}{Name of the object to save in quotes}

\item{dir}{Directory path}

\item{folder}{\code{TRUE} if looking for folder names}
}
\value{
Object name of the most recent file
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
look into the folder and find the version of the file
with the most recent name
}
\keyword{internal}
