% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_clean_chron_temp.R
\name{util_clean_chron_temp}
\alias{util_clean_chron_temp}
\title{Delete all temporary files for age-depth modelling}
\usage{
util_clean_chron_temp(dir)
}
\arguments{
\item{dir}{Path to the data storage folder}
}
\description{
Delete all temporary files for age-depth modelling
}
