% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_compare_files.R
\name{util_compare_files}
\alias{util_compare_files}
\title{Compare two datasets}
\usage{
util_compare_files(
  file_name_a,
  file_name_b,
  file_a_envir = NULL,
  file_b_envir = NULL
)
}
\arguments{
\item{file_name_a}{The name of the object to compare in quotes}

\item{file_name_b}{The name of the object to compare in quotes}

\item{file_a_envir}{The environment where is the \code{file_a} present}

\item{file_b_envir}{The environment where is the \code{file_b} present}
}
\value{
Logical. Are those file the same?
}
\description{
Compare the two files and return if same
}
