% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_confirm.R
\name{util_confirm}
\alias{util_confirm}
\title{Open menu to confirm TRUE/FALSE}
\usage{
util_confirm(msg = "", default_value = TRUE)
}
\arguments{
\item{msg}{Message to be shown with the menu}

\item{default_value}{Default value of answer}
}
\value{
return logical
}
\description{
Open menu to confirm TRUE/FALSE
}
\keyword{internal}
