% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_confirm_based_on_presence.R
\name{util_confirm_based_on_presence}
\alias{util_confirm_based_on_presence}
\title{Confirm choice based on the presence of previous data}
\usage{
util_confirm_based_on_presence(dir, file_name, msg = "")
}
\arguments{
\item{dir}{Path to the folder, where data should be stored}

\item{file_name}{Name of the file to look upon}

\item{msg}{Message to be placed with the confirmation question}
}
\description{
Confirm choice based on the presence of previous data
}
