% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_confirm_global_setting.R
\name{util_confirm_global_setting}
\alias{util_confirm_global_setting}
\title{Confirm the general setting}
\usage{
util_confirm_global_setting(setting_name)
}
\arguments{
\item{setting_name}{Name of the setting variable}
}
\description{
Offer the user to temporarily change one logical variable
}
