% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_extract_config_data.R
\name{util_extract_config_data}
\alias{util_extract_config_data}
\title{Extract Config information from Global environment}
\usage{
util_extract_config_data()
}
\value{
List with all the important data-related settings
}
\description{
Gather information about current setting config.
}
