% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_extract_var_safe.R
\name{util_extract_var_safe}
\alias{util_extract_var_safe}
\title{Helper function to extract data safely}
\usage{
util_extract_var_safe(var, dataset)
}
\arguments{
\item{var}{Name of the variable}

\item{dataset}{list of lists}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
