% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_missing_seq_names.R
\name{util_get_missing_seq_names}
\alias{util_get_missing_seq_names}
\title{Get names of missing datasets}
\usage{
util_get_missing_seq_names(dir, name_vector)
}
\arguments{
\item{dir}{Character. Path to the folder with files}

\item{name_vector}{Vector with names of datasets.}
}
\description{
Get names of missing datasets
}
