% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_load_chron_crashfile.R
\name{util_load_chron_crashfile}
\alias{util_load_chron_crashfile}
\title{Load last Crash file or create dummy if not present}
\usage{
util_load_chron_crashfile(dir)
}
\arguments{
\item{dir}{Path to the data storage folder}
}
\description{
Load last Crash file or create dummy if not present
}
