% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_make_chron_crashfile.R
\name{util_make_chron_crashfile}
\alias{util_make_chron_crashfile}
\title{Make Bchron Crash file}
\usage{
util_make_chron_crashfile(dir)
}
\arguments{
\item{dir}{Path to the directory}
}
\description{
Make Bchron Crash file
}
