% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_make_datastorage_folders.R
\name{util_make_datastorage_folders}
\alias{util_make_datastorage_folders}
\title{Test presence and create folders if necessary}
\usage{
util_make_datastorage_folders(dir)
}
\arguments{
\item{dir}{Path to the selected directory}
}
\description{
Test presence and create folders if necessary
}
