% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_make_fig_dir.R
\name{util_make_fig_dir}
\alias{util_make_fig_dir}
\title{Create folders for figure outputs}
\usage{
util_make_fig_dir(dir, region_vector, sel_date)
}
\arguments{
\item{dir}{Path to the directory}

\item{region_vector}{Vector with names of folders to create}

\item{sel_date}{Date to be used}
}
\description{
Create a series of folders for all \code{region_vector} with
\code{sel_date} inside of \code{dir}
}
\keyword{internal}
