% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_open_dir.R
\name{util_open_dir}
\alias{util_open_dir}
\title{Open selected directory}
\usage{
util_open_dir(dir)
}
\arguments{
\item{dir}{Path to the directory}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
