% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_open_dir_if_not.R
\name{util_open_dir_if_not}
\alias{util_open_dir_if_not}
\title{Testing any criteria and open folder with msg if not true}
\usage{
util_open_dir_if_not(..., env = parent.frame(), dir, msg)
}
\arguments{
\item{...}{Unnamed expressions that describe the conditions to be tested}

\item{env}{The environment in which to evaluate the test}

\item{dir}{Path to the selected folder}

\item{msg}{Error message to be printed}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Test a selected criteria an open selected directory if not true
}
