% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_paste_as_vector.R
\name{util_paste_as_vector}
\alias{util_paste_as_vector}
\title{Paste any list of values as a vector}
\usage{
util_paste_as_vector(var_list, sep = "'")
}
\arguments{
\item{var_list}{List of values}

\item{sep}{Separator which should be used to wrap each object in
the \code{var_list}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
