% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_replace_null_with_na.R
\name{util_replace_null_with_na}
\alias{util_replace_null_with_na}
\title{Replace NULL with N}
\usage{
util_replace_null_with_na(x)
}
\arguments{
\item{x}{Any R object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Helper function. If \code{x} is \code{NULL} then replace with \code{NA}
}
\keyword{internal}
