% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_save_if_latests.R
\name{util_save_if_latests}
\alias{util_save_if_latests}
\title{Save the newest version of dataset}
\usage{
util_save_if_latests(
  file_name,
  dir,
  system_setting_file = "current_setting",
  prefered_format = c("rds", "csv"),
  compress = TRUE
)
}
\arguments{
\item{file_name}{Name of the object to save in quotes}

\item{dir}{Directory path}

\item{system_setting_file}{Character. Name of the object in global environments
which contains a description of setting}

\item{prefered_format}{format to save as: "rda" or "csv"}

\item{compress}{Should \code{gz} compression be applied?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Look into the folder and find the version of the file with
the most recent name. Compare the last saved file and selected file and
save file if something changed since recent
}
\keyword{internal}
