% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stop_if_not.R
\name{util_stop_if_not}
\alias{util_stop_if_not}
\title{Testing any criteria and either stop or produce msg}
\usage{
util_stop_if_not(..., env = parent.frame(), false_msg = NULL, true_msg = NULL)
}
\arguments{
\item{...}{Unnamed expressions that describe the conditions to be tested}

\item{env}{The environment in which to evaluate the test}

\item{false_msg}{Error message to be printed when stopping}

\item{true_msg}{Comment to be printed when positive result of test}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Test a selected criteria an output comment or stop
}
\keyword{internal}
