% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stop_quietly.R
\name{util_stop_quietly}
\alias{util_stop_quietly}
\title{Stop the script}
\usage{
util_stop_quietly()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Stop the script without producing error
}
\keyword{internal}
