% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_recalibrate_in_batches.R
\name{chron_recalibrate_in_batches}
\alias{chron_recalibrate_in_batches}
\title{Recalibrate the selected sequences in defined batches using parallel
computation}
\usage{
chron_recalibrate_in_batches(
  data_source_chron,
  batch_size = 1,
  time_per_sequence = 120,
  dir,
  n_iterations = 10000,
  n_burn = 2000,
  n_thin = 8,
  number_of_cores = 1,
  set_seed = 1234,
  maximum_number_of_loops = 3
)
}
\arguments{
\item{data_source_chron}{Data.frame containing \code{dataset_id} and \code{chron_control_format}}

\item{batch_size}{Numeric. Number of individual sequences re-calibrate in a single
batch using parallel computation}

\item{time_per_sequence}{Time (in sec) dedicated for each sequence to estimate
age-depth model. If it takes computer longer that selected value, estimation
is considered as unsuccessful and skipped. The time value is multiplied by
\code{iteration_multiplier} as more iteration required more time. Time for whole
batch is calculated as \code{time_per_sequence} multiplied by \code{iteration_multiplier}
multiplied by the number of sequences per batch (which is estimated based on
\code{number_of_cores})}

\item{dir}{Character. Path to the data storage folder}

\item{n_iterations}{Numeric. The number of iterations used by Bchron}

\item{n_burn}{Numeric. The number of starting iterations to discard used by Bchron}

\item{n_thin}{Numeric. The step size for every iteration to keep beyond
the burnin used by Bchron}

\item{number_of_cores}{Numeric. Number of CPU cores to use in parallel computation}

\item{set_seed}{Numeric. User-defined seed for randomisations}

\item{maximum_number_of_loops}{Numeric. Number of tries each batch should be considered
before skipping it}
}
\value{
Vector with names of sequences which failed to estimate
}
\description{
Recalibrate the selected sequences in defined batches using parallel
computation
}
