% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron_recalibrate_individual.R
\name{chron_recalibrate_individual}
\alias{chron_recalibrate_individual}
\title{Recalibrate age-depth models for individual sequences}
\usage{
chron_recalibrate_individual(
  data_source_chron,
  n_iterations = 10000,
  n_burn = 2000,
  n_thin = 8,
  time_to_stop = 600,
  dir
)
}
\arguments{
\item{data_source_chron}{Data.frame containing \code{dataset_id} and \code{chron_control_format}}

\item{n_iterations}{Numeric. The number of iterations used by Bchron}

\item{n_burn}{Numeric. The number of starting iterations to discard used by Bchron}

\item{n_thin}{Numeric. The step size for every iteration to keep beyond
the burnin used by Bchron}

\item{time_to_stop}{Numeric. Time to wait until saving the sequence's \code{dataset_id} into
\code{Crash_file}as unsuccessfully, due to freeze of calculation}

\item{dir}{Character. Path to the data storage folder}
}
\description{
Recalibrate age-depth models for individual sequences
}
