% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_detect_duplicates.R
\name{proc_detect_duplicates}
\alias{proc_detect_duplicates}
\title{Detect and exclude duplicates}
\usage{
proc_detect_duplicates(
  data_source,
  source_var = "source_of_data",
  n_subgroups = 1,
  maximal_distance = 5
)
}
\arguments{
\item{data_source}{Data.frame with records}

\item{source_var}{Character. Name of a column that indicates source of data}

\item{n_subgroups}{Number of subgroups to split the data based on the
geography. Records within subgroup will be tested only if subgroup
contains records form different \code{source_var}}

\item{maximal_distance}{Maximal Euclidean distance between two records to
consider them as material for comparison}
}
\value{
Data.frame with the suggested duplicates of records
\code{distance} = Euclidean distance between records
\code{similarity} = mean number of the exactly same columns
}
\description{
Function will split records based on their geographic
location into \code{n_subgroups}. If subgroup contains data from multiple sources,
Euclidean \code{distance} between all records combination will be calculated.
Only combination with the distance up to \code{maximal_distance} will be kept.
Next, selected columns will be compared between records and the average number of
the exact same columns is returned \code{similarity}.
}
