% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_get_merged_dataset.R
\name{proc_get_merged_dataset}
\alias{proc_get_merged_dataset}
\title{Merge data from Neotoma and Private source}
\usage{
proc_get_merged_dataset(data_storage_path, private_data = FALSE)
}
\arguments{
\item{data_storage_path}{Path to the data storage folder}

\item{private_data}{Logical value if to include private source of data}
}
\description{
Merge data from Neotoma and Private source
}
