% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_download_sequences.R
\name{proc_neo_download_sequences}
\alias{proc_neo_download_sequences}
\title{Download all sequences from Neotoma}
\usage{
proc_neo_download_sequences(allds, n_tries = 10)
}
\arguments{
\item{allds}{Data.frame with all dataset IDs}

\item{n_tries}{Number of tries to download each dataset}
}
\description{
Download all sequences from Neotoma
}
