% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_neo_get_all_neotoma_datasets.R
\name{proc_neo_get_all_neotoma_datasets}
\alias{proc_neo_get_all_neotoma_datasets}
\title{Download the list of all Neotoma datasets}
\usage{
proc_neo_get_all_neotoma_datasets(
  dataset_type,
  long_min,
  long_max,
  lat_min,
  lat_max
)
}
\arguments{
\item{dataset_type}{Type of dataset to use}

\item{long_min}{Limit for the smallest longitude}

\item{long_max}{Limit for the largest longitude}

\item{lat_min}{Limit for the smallest latitude}

\item{lat_max}{Limit for the largest latitude}
}
\description{
Download all datasets from Neotoma and subset only those for
selected \code{dataset_type} and within selected geographical limits
}
