% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_save_references.R
\name{proc_save_references}
\alias{proc_save_references}
\title{Save all the reference tables of the data assembly}
\usage{
proc_save_references(
  data_source,
  project_database,
  user_sel_variables = c(),
  dir
)
}
\arguments{
\item{data_source}{Data.frame with \code{dataset_id} and other variables}

\item{project_database}{Project dataset database}

\item{user_sel_variables}{Vector with variables, which have to be present in the
final data assembly}

\item{dir}{Path to the data storage folder}
}
\description{
Save all the reference tables of the data assembly
}
